Avoid using local registry when adding imported packages
Flexmi uses the local "resourceSet().getPackageRegistry()" as a
namespace registry to look for EClass names when doing fuzzy matching.
Any extra package added to that registry would worse loading times (by a
lot if several packages are added / they are big), and also strange
errors can appear (e.g. if using "package" as tag name for EPackages,
and also adding UML to the local registry, then the "package" tag would
be matched with UML's Package class).

So, when loading Flexmi models programmatically, any imported packages
of interest must be added to the global EPackage.Registry instance.

TODO: maybe improve this so that using the local registry is not an
issue. I think this should be at least mentioned in the Flexmi docs.